Farm App
The project is about creating a Java application for a large farm to manage the cows, their treatments, and medication, as well as the vet staff working on the farm. It aims to maintain information about the cows, their illnesses, and also the
history of the treatment of their illnesses.
Advanced Java programming, object-oriented programming, exception handling, multithreading, I/O Streams,
graphical user interface, Java Database Connectivity (MySQL), Serialization, and Security (MD5 generation).